emotional-ai-chatbot-bak/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.8.3
grpcio==1.60.0
grpcio-tools==1.60.0
